requirements.txt
django-autoslug==1.4.1
